containers/images/hawkular-agent-mock/requirements.txt
locust
hawkular-client
benchpy
numpy